Refactored selection tools so as to not duplicate them between the scatter and volume viewers. In addition, selections with multiple datasets now work properly (except for point selections, which use the selected layer if it is a Data layer, or the first Data layer if not).
