+++ glue_vispy_viewers/common/toolbar.py
@@ -8,7 +8,7 @@ from __future__ import absolute_import, division, print_function
 
 from glue.viewers.common.qt.toolbar import BasicToolbar
 
-from .tools import VispyMouseMode
+from .selection_tools import VispyMouseMode
 
 
 class VispyViewerToolbar(BasicToolbar):
